seqAlleleFreq {SeqArray}	R Documentation

Get Allele Frequencies

Description

Calculates the allele frequencies.

Usage

seqAlleleFreq(gdsfile, ref.allele=0L,
    parallel=getOption("seqarray.parallel", FALSE))

Arguments

gdsfile
a SeqVarGDSClass object

ref.allele
NULL, a single numeric value, a numeric vector or a character vector; see Value

parallel
FALSE (serial processing), TRUE (parallel processing) or other value; parallel is passed to the argument cl in seqParallel, see seqParallel for more details.

Value

If ref.allele=NULL, the function returns a list of allele frequencies according to all allele per site. If ref.allele is a single numeric value (like 0L), it returns a numeric vector for the specified alleles (0L for the reference allele, 1L for the first alternative allele, etc). If ref.allele is a numeric vector, ref.allele specifies each allele per site. If ref.allele is a character vector, ref.allele specifies the desired allele for each site (e.g., ancestral allele for the derived allele frequency).

Author(s)

Xiuwen Zheng

See Also

seqNumAllele, seqMissing, seqParallel

Examples

# the GDS file
(gds.fn <- seqExampleFileName("gds"))

# display
f <- seqOpen(gds.fn)

# return a list
head(seqAlleleFreq(f, NULL))

# return a numeric vector
summary(seqAlleleFreq(f, 0L))

# return a numeric vector, AA is ancestral allele
AA <- toupper(seqGetData(f, "annotation/info/AA")$data)
summary(seqAlleleFreq(f, AA))

# close the GDS file
seqClose(f)

[Package SeqArray version 1.10.1 Index]