seqAsVCF {SeqArray}	R Documentation

VariantAnnotation objects

Description

Imports the VariantAnnotation package and export the generic methods associated with SeqVarGDSClass class.

Usage

seqAsVCFInit()
seqAsVCF(gdsfile, ...)

Arguments

gdsfile
a SeqVarGDSClass object

...
optional arguments to asVCF

Details

The generic methods ref, alt, qual, filt, fixed, header, info, geno, asVCF, rowRanges and colData are exported and associated with a SeqVarGDSClass object after calling seqAsVCFInit() or seqAsVCF().

Value

None or a CollapsedVCF object.

Author(s)

Xiuwen Zheng

See Also

asVCF,SeqVarGDSClass-method

Examples

gds <- seqOpen(seqExampleFileName("gds"))

## Not run: 
seqAsVCFInit()  ## enable exporting the generic methods
seqAsVCF(gds)

## End(Not run)

seqClose(gds)

[Package SeqArray version 1.17.1 Index]