seqExampleFileName {SeqArray}	R Documentation

Example files

Description

The example files of VCF and GDS format.

Usage

seqExampleFileName(type=c("gds", "vcf", "KG_Phase1"))

Arguments

type
either "gds" or "vcf"

Details

The SeqArray GDS file was created from a subset of VCF data of the 1000 Genomes Phase 1 Project.

Value

Return the file name of a VCF file shipped with the package if type = "vcf", or the file name of a GDS file if type = "gds".

Author(s)

Xiuwen Zheng

Examples

seqExampleFileName("gds")

seqExampleFileName("vcf")

seqExampleFileName("KG_Phase1")

[Package SeqArray version 1.12.9 Index]