seqMerge {SeqArray}	R Documentation

Merge Multiple Sequence GDS Files

Description

Merges multiple sequence GDS files.

Usage

seqMerge(gds.fn, out.fn, storage.option=seqStorage.Option(), verbose=TRUE)

Arguments

gds.fn
the file names of multiple GDS files

out.fn
the output file name

storage.option
specify the storage and compression options, by default seqStorage.Option

verbose
if TRUE, show information

Details

The current implementation of seqMerge extracts and merges the genotypic data only without any annotation. Users can specify multiple VCF files in seqVCF2GDS to export a single GDS file.

Value

None.

Author(s)

Xiuwen Zheng

See Also

seqVCF2GDS

[Package SeqArray version 1.10.2 Index]