seqMissing {SeqArray}	R Documentation

Missing genotype percentage

Description

Calculates the missing rates per variant or per sample.

Usage

seqMissing(gdsfile, per.variant=TRUE,
    parallel=getOption("seqarray.parallel", FALSE))

Arguments

gdsfile
a SeqVarGDSClass object

per.variant
missing rate per variant if TRUE, or missing rate per sample if FALSE

parallel
FALSE (serial processing), TRUE (parallel processing) or other value; parallel is passed to the argument cl in seqParallel, see seqParallel for more details.

Value

A vector of missing rates.

Author(s)

Xiuwen Zheng

See Also

seqAlleleFreq, seqNumAllele, seqParallel

Examples

# the GDS file
(gds.fn <- seqExampleFileName("gds"))

# display
(f <- seqOpen(gds.fn))

summary(seqMissing(f, TRUE))

summary(seqMissing(f, FALSE))

# close the GDS file
seqClose(f)

[Package SeqArray version 1.10.1 Index]