seqVCF2GDS {SeqArray}	R Documentation

Reformat VCF Files

Description

Reformats Variant Call Format (VCF) files.

Usage

seqVCF2GDS(vcf.fn, out.fn, header=NULL, genotype.var.name="GT",
    genotype.storage=c("bit2", "bit4", "bit8"),
    storage.option=seqStorage.Option(),
    info.import=NULL, fmt.import=NULL, ignore.chr.prefix="chr",
    reference=NULL, optimize=TRUE, raise.error=TRUE, verbose=TRUE)

Arguments

vcf.fn
the file name(s) of VCF format

out.fn
the file name of output GDS file

header
if NULL, header is set to be seqVCF.Header(vcf.fn)

genotype.var.name
the ID for genotypic data in the FORMAT column; "GT" by default, VCFv4.0

genotype.storage
"bit2" by default; with respect to the compression size and access speed, "bit2" is the most efficient when most of variants are biallelic.

storage.option
specify the storage and compression options, by default seqStorage.Option, see details

info.import
characters, the variable name(s) in the INFO field for import; or NULL for all variables

fmt.import
characters, the variable name(s) in the FORMAT field for import; or NULL for all variables

ignore.chr.prefix
a vector of character, indicating the prefix of chromosome which should be ignored, like "chr"; it is not case-sensitive

reference
genome reference, like "hg19", "GRCh37"

optimize
if TRUE, optimize the access efficiency by calling cleanup.gds

raise.error
TRUE: throw an error if numeric conversion fails; FALSE: get missing value if numeric conversion fails

verbose
if TRUE, show information

Details

GDS – Genomic Data Structures used for storing genetic array-oriented data, and the file format defined in the gdsfmt package.

VCF – The Variant Call Format (VCF), which is a generic format for storing DNA polymorphism data such as SNPs, insertions, deletions and structural variants, together with rich annotations.

If there are more than one files in vcf.fn, seqVCF2GDS will merge all VCF files together if they contain the same samples. It is useful to merge genomic variants if VCF data are divided by chromosomes.

The real numbers in the VCF file(s) are stored in 32-bit floating-point format by default. Users can set seqStorage.Option(float.mode="float64") to switch to 64-bit floating point format. Or packed real numbers can be adopted by setting seqStorage.Option(float.mode="packedreal16:scale=0.0001").

By default, the compression method is "ZIP_RA" (zlib algorithm with default compression level + multiple independent data blocks). Users can maximize the compression ratio by storage.option=seqStorage.Option("ZIP_RA.max"). LZ4 (an extremely fast compression algorithm) is an option via storage.option=seqStorage.Option("LZ4_RA").

Value

Return the file name of GDS format with an absolute path.

Author(s)

Xiuwen Zheng

References

The variant call format and VCFtools. Danecek P, Auton A, Abecasis G, Albers CA, Banks E, DePristo MA, Handsaker RE, Lunter G, Marth GT, Sherry ST, McVean G, Durbin R; 1000 Genomes Project Analysis Group. Bioinformatics. 2011 Aug 1;27(15):2156-8. Epub 2011 Jun 7.

http://cyan4973.github.io/lz4/

See Also

seqVCF.Header, seqStorage.Option, seqGDS2VCF

Examples

# the VCF file
vcf.fn <- seqExampleFileName("vcf")

# convert
seqVCF2GDS(vcf.fn, "tmp.gds")

# display
(f <- seqOpen("tmp.gds"))
seqClose(f)

# convert without the INFO fields
seqVCF2GDS(vcf.fn, "tmp.gds", info.import=character(0))

# display
(f <- seqOpen("tmp.gds"))
seqClose(f)

# convert without the INFO and FORMAT fields
seqVCF2GDS(vcf.fn, "tmp.gds",
    info.import=character(0), fmt.import=character(0))

# display
(f <- seqOpen("tmp.gds"))
seqClose(f)

# delete the temporary file
unlink("tmp.gds")

[Package SeqArray version 1.10.2 Index]